feat: add default font-size
I've used fluid typography conservatively where it makes sense. A base
font size of 16px on mobile devices and not increasing it makes sense
because we can fit more characters on a single line compared to a page
with a base font size of 17px or some higher value.

This base font size will also affect the size of all the other elements
on the page which ends up giving us fluid design with a single line of
CSS rather than convoluted sets of clamp() functions for all possible
scenarios. The clamp() function and the values used in this commit are
pretty similar to the one used on https://web.dev as on the date of this
commit.