feat: add fluid spacing for lists and dd
This is another conservative use of fluid typography. On a desktop with
ample real estate and resolution, we want increasing padding for <ol>
and <ul> but on a mobile phone, the same amount of padding is waste of
space and makes more likely that list elements might overflow. These
clamp() functions reduce the padding the list elements on mobile but
increase it to a more pleasant size on desktop.