feat: add rules for html tag
These rules for the <html> tag are inspired from a lot of different
sources. One of the primary resources behind this commit was this blog
post

https://www.joshwcomeau.com/css/custom-css-reset/

The motivation to segregate default rules, screen specific rules, and
print specific rules came from the CSS on https://seirdy.one as on the
date of this commit. The `text-size-adjust` property rule also came from
his website stylesheet. This property is not supported on all browsers.

I think it's helpful to be able to override the border-box and have the
override inherit to descendant elements as well which is why `html {}`
has the border-box rule rather than `* {}`.

I won't remove the default margin from all elements. Repairing the
damage done by this apparently attractive rule is hard to anticipate and
fix.

The line-height has been set to 1.6 by default which meets the "WCAG
1.4.12: Text Spacing" recommendation about line height.

The color-scheme shouldn't be a part of this stylesheet but it has been
included in case

<meta name="color-scheme" content="light dark">

isn't included in the HTML page.