feat: add rules for headings
The size of headings will apply to both print and display for
consistency. However, the color of headings on display has lesser
contrast than body text to try and reduce possibilities of halation. The
resulting contrast still meets APCA guidelines.

The line height of headings is higher on screen than on print.