feat: normalize fonts for various tags
One of the objectives behind this stylesheet is to distinguish tags
based on not just colors but also something textual, even if it's just
borders or font weight. <kbd>, <var> and <dt> have been assigned a
font-weight of 600.

<address> and <time> have been assigned monospace fonts to eliminate
possibilities of character ambiguity.

Instead of using italic fonts, which are sometimes not truly italic and
just slanted, we'll use the increased font weights to distinguish
<address>, <dfn>, and <var>. However, <del>, <ins>, and <q> still retain
italicized fonts.