feat: remove list mark in lists with explicit role
If a list element has been explicitly assigned a role of a list, it's
expected to be used in a scenario where its list markers and padding has
been removed for some other purpose. The most common use case I've come
across for this is in creating navigation menu links at the start or at
the end of a page using either `display: flex`, `inline-block` for
better backwards compatibility, or `inline-flex`.