feat: add width, height, and padding to textarea
This feels somewhat of a compromize. For one thing, <textarea> doesn't
seem to respect max-width on all browsers and overflows horizontally,
which messes things up, which is why I've add a width of 100%. However,
this means ignoring the value of cols.

If rows attribute isn't mentioned, a default value of 15ex has been
added.