feat: add flexbox layout to lists in nav
If a list has been explicitly assigned the role of "list", it is most
likely intended to be used inside nav without its list markers. A
flexbox layout in this case works well and wraps nav links as expected.

feat: add padding to nav links

nav links are tap targets and should be easily clickable on touch
devices